Fix release builds for musl targets
The `libpg_query.a` static archive fails to link during a release build
for musl targets with errors like: "undefined reference to
`_longjmp_chk'".

This appears to be caused by the fact that the static archive is
compiled with the default toolchain, and so has references to symbols
provided by glibc (but not musl libc).

This change switches to use the `musl-gcc` wrapper as `CC` when the
target triple ends with `-musl`.